+++ final/src/datasources/cats-api.js
@@ -1,6 +1,6 @@
 import { RESTDataSource } from "@apollo/datasource-rest";
 
-export class CatstronautsAPI extends RESTDataSource {
+export class CatsAPI extends RESTDataSource {
   constructor() {
     super();
     // the Catstronauts catalog is hosted on this server
