+++ server/package.json
@@ -3,18 +3,24 @@
   "version": "1.0.0",
   "description": "Back-end GraphQL API for GraphQL Summit 2022 workshop",
   "main": "src/index.js",
+  "type": "module",
   "scripts": {
     "start": "node src/index",
     "dev": "nodemon src/index -e js,graphql"
   },
   "dependencies": {
-    "apollo-datasource-rest": "^0.11.0",
-    "apollo-server": "^3.10.2",
-    "graphql": "^15.8.0"
+    "@apollo/datasource-rest": "4.2.0",
+    "@apollo/server": "^4.0.0-alpha.12",
+    "graphql": "^16.5.0",
+    "graphql-tag": "^2.12.6"
   },
   "devDependencies": {
+    "@graphql-tools/mock": "^8.7.6",
+    "@graphql-tools/schema": "^9.0.4",
+    "@types/node": "^17.0.21",
     "dotenv": "^8.2.0",
-    "nodemon": "^2.0.4"
+    "nodemon": "^2.0.4",
+    "typescript": "^4.8.3"
   },
   "engines": {
     "npm": ">=6.0.0",
