--- server4/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "catstronauts-server",
-  "version": "1.0.0",
-  "description": "Back-end GraphQL API for GraphQL Summit 2022 workshop",
-  "main": "src/index.js",
-  "scripts": {
-    "start": "node src/index",
-    "dev": "nodemon src/index -e js,graphql"
-  },
-  "dependencies": {
-    "@apollo/server": "^4.0.0-alpha.12",
-    "apollo-datasource-rest": "^0.11.0",
-    "graphql": "^16.5.0",
-    "graphql-tag": "^2.12.6"
-  },
-  "devDependencies": {
-    "dotenv": "^8.2.0",
-    "nodemon": "^2.0.4"
-  },
-  "engines": {
-    "npm": ">=6.0.0",
-    "node": ">=14.0.0 <17.0.0"
-  },
-  "license": "MIT",
-  "private": true
-}
