--- server4/src/datasources/catstronauts-api.js
@@ -1,39 +1,0 @@
-const { RESTDataSource } = require('apollo-datasource-rest');
-
-class CatstronautsAPI extends RESTDataSource {
-  constructor() {
-    super();
-    // the Catstronauts catalog is hosted on this server
-    this.baseURL = 'https://odyssey-lift-off-rest-api.herokuapp.com/';
-  }
-
-  getAllTracks() {
-    return this.get('tracks');
-  }
-
-  getAuthor(authorId) {
-    return this.get(`author/${authorId}`);
-  }
-
-  getTrack(trackId) {
-    return this.get(`track/${trackId}`);
-  }
-
-  getTrackModules(trackId) {
-    return this.get(`track/${trackId}/modules`);
-  }
-
-  getModule(moduleId) {
-    return this.get(`module/${moduleId}`);
-  }
-
-  incrementTrackViews(trackId) {
-    return this.patch(`track/${trackId}/numberOfViews`);
-  }
-
-  incrementTrackLikes(trackId) {
-    return this.patch(`track/${trackId}/numberOfLikes`);
-  }
-}
-
-module.exports = CatstronautsAPI;
