--- server4/src/resolvers.js
@@ -1,1 +1,0 @@
-// TODO: Resolvers go here
