--- server4/src/schema.graphql
@@ -1,7 +1,0 @@
-type Query {
-  """
-  This is just a placeholder for the starter code.
-  Delete it once you're able to run a test query in Sandbox!
-  """
-  deleteMe: String
-}
