--- server4_module/src/index.js
@@ -1,25 +1,0 @@
-import { readFileSync } from "fs";
-import { ApolloServer } from "@apollo/server";
-import { startStandaloneServer } from "@apollo/server/standalone";
-import gql from "graphql-tag";
-
-import { addMocksToSchema } from "@graphql-tools/mock";
-import { makeExecutableSchema } from "@graphql-tools/schema";
-import { mocks } from "./mocks.js";
-
-const typeDefs = gql(
-  readFileSync("./src/schema.graphql", { encoding: "utf-8" })
-);
-
-const server = new ApolloServer({
-  schema: addMocksToSchema({
-    schema: makeExecutableSchema({ typeDefs }),
-    mocks,
-  }),
-});
-
-const { url } = await startStandaloneServer(server, {
-  listen: { port: process.env.PORT || 4000 },
-});
-
-console.log(`🚀  Server ready at ${url}`);
